backup_20250531_140359/pyproject.toml
# pyproject.toml
[project]
name = "dbcreds"
version = "2.0.0"
description = "Professional database credentials management with security and team collaboration in mind"
readme = "README.md"
authors = [
    {name = "Your Company", email = "dev@yourcompany.com"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Database",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]
requires-python = ">=3.8"
dependencies = [
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "loguru>=0.7.0",
    "keyring>=24.0.0",
    "psycopg2-binary>=2.9.0",
    "sqlalchemy>=2.0.0",
    "python-multipart>=0.0.6",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "jinja2>=3.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "aiofiles>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
    "mkdocstrings[python]>=0.23.0",
]
mysql = ["mysqlclient>=2.2.0"]
oracle = ["oracledb>=1.4.0"]
mssql = ["pyodbc>=5.0.0"]

[project.scripts]
dbcreds = "dbcreds.cli:app"
dbcreds-server = "dbcreds.web.__main__:main"
dbcreds-migrate = "dbcreds.migrate:app"

[project.urls]
Homepage = "https://github.com/yourcompany/dbcreds"
Documentation = "https://yourcompany.github.io/dbcreds"
Repository = "https://github.com/yourcompany/dbcreds"
Issues = "https://github.com/yourcompany/dbcreds/issues"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["dbcreds", "dbcreds.backends", "dbcreds.core", "dbcreds.web", "dbcreds.utils"]

[tool.ruff]
line-length = 120
target-version = "py38"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

# uv configuration
[tool.uv]
managed = true
dev-dependencies = [
    "ipython>=8.12.0",
    "ipdb>=0.13.13",
]